app2-functions/requirements.txt
azure-functions==1.18.0
azure-functions-durable==1.2.9
langchain==0.1.10
langchain-openai==0.0.8
openai==1.12.0
